-Improved DbFixture so it generates the DB only once. -Cleaned up some tests that left garbage after execution -Introduced IBasicQuery and cleaned up the queries in Application -Removed the unnecessary reference to IServiceProvider that was all over handlers

+++ tests/Infrastructure.Tests/Data/Helpers/DbContextFixture.cs
@@ -14,24 +14,20 @@ public class DbContextFixture : IAsyncLifetime
         DbContext = dbContextFactory?.CreateDbContext() ?? throw new ArgumentNullException(nameof(dbContextFactory));
 
         _outputAccessor = outputAccessor ?? throw new ArgumentNullException(nameof(outputAccessor));
-        Seeder = seeder ?? throw new ArgumentNullException(nameof(seeder));
+        _seeder = seeder ?? throw new ArgumentNullException(nameof(seeder));
         Configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
-        //try
-        //{
-            DbContext.Database.EnsureDeleted();
-        //}
-        //catch
-        //{ }
-        DbContext.Database.Migrate();
-    }    
 
-    readonly ITestOutputHelperAccessor _outputAccessor;
-    readonly IDbSeeder Seeder;
+        if (!_dbMigrated)
+        {
+            DbContext.Database.EnsureDeleted();
+            DbContext.Database.Migrate();
+
+            _dbMigrated = true;
+        }
+    }    
 
     public ITestOutputHelper Output => _outputAccessor.Output!;
 
-
-    protected bool SkipInsertTestData { get; set; } = false;
     [Obsolete("This is a hack to check the database data if tests don't run successfully.")]
     public bool SkipDeletingDatabase { get; set; }
 
@@ -45,15 +41,32 @@ public class DbContextFixture : IAsyncLifetime
             return DbContext.Database.EnsureDeletedAsync();
 #pragma warning restore CS0618 // Type or member is obsolete
 
-        
+        DbContext.Dispose();
+
         return Task.CompletedTask;
     }
 
     public async Task InitializeAsync()
     {
-        if (SkipInsertTestData)
+        if (_dbSeeded)
             return;
+        
+        _dbSeeded = true;
 
-        await Seeder.SeedAsync();
+        await _seeder.SeedAsync();
+
+        DbContext.Database.ExecuteSqlRaw("alter fulltext index on dbo.Videos start full population");
+        DbContext.Database.ExecuteSqlRaw("alter fulltext index on dbo.Playlists start full population");
+        DbContext.Database.ExecuteSqlRaw("alter fulltext index on dbo.Transcripts start full population");
+        DbContext.Database.ExecuteSqlRaw("alter fulltext index on dbo.Artifacts start full population");
+
+        // TODO: I should do what's described here: https://stackoverflow.com/questions/2727911/how-can-i-know-when-sql-full-text-index-population-is-finished
+        // For now this seems to work on my machine.
+        await Task.Delay(1500);
     }
+
+    static bool _dbMigrated = false;
+    static bool _dbSeeded = false;
+    readonly ITestOutputHelperAccessor _outputAccessor;
+    readonly IDbSeeder _seeder;
 }
